@@ -10,36 +10,44 @@
     "node": ">=10"
   },
   "devDependencies": {
-    "@mapbox/appropriate-images": "^3.0.0",
+    "@babel/core": "^7.9.6",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-flow": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
+    "@mapbox/appropriate-images": "^4.0.0",
     "@mapbox/appropriate-images-react": "^1.0.0",
     "@mapbox/batfish": "1.9.8",
     "@mapbox/copyeditor": "^0.7.0",
-    "@mapbox/dr-ui": "^0.27.0",
-    "@mapbox/eslint-config-docs": "^0.3.0",
-    "@mapbox/eslint-config-mapbox": "^2.0.1",
+    "@mapbox/dr-ui": "^0.29.1",
+    "@mapbox/eslint-config-docs": "^0.6.0",
+    "@mapbox/eslint-config-mapbox": "^3.0.0",
+    "@mapbox/mapbox-gl-style-spec": "^13.15.0",
     "@mapbox/mapbox-gl-supported": "^1.4.0",
     "@mapbox/mbx-assembly": "^0.29.0",
-    "@mapbox/mr-ui": "^0.7.4",
+    "@mapbox/mr-ui": "^0.9.1",
     "@mapbox/prettier-config-docs": "^0.1.0",
-    "@mapbox/remark-lint-link-text": "^0.4.5",
-    "@mapbox/remark-lint-mapbox": "^0.10.2",
-    "@sentry/browser": "^5.13.0",
-    "babel-eslint": "^10.0.3",
-    "core-js": "^3.1.4",
+    "@mapbox/remark-config-docs": "0.5.0",
+    "@mapbox/remark-lint-link-text": "^0.5.0",
+    "@mapbox/remark-lint-mapbox": "^0.12.1",
+    "@sentry/browser": "^5.15.5",
+    "babel-eslint": "^10.1.0",
     "documentation": "~12.1.0",
-    "eslint": "^6.1.0",
-    "eslint-config-mourner": "^3.0.0",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-plugin-es": "^2.0.0",
-    "eslint-plugin-html": "^6.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-node": "^10.0.0",
+    "downshift": "^5.4.2",
+    "eslint": "^7.4.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-es": "^3.0.1",
+    "eslint-plugin-html": "^6.0.2",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-xss": "^0.1.9",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-xss": "^0.1.10",
     "execa": "^3.2.0",
     "github-slugger": "1.2.1",
-    "husky": "^4.0.9",
+    "husky": "^4.2.3",
+    "jest": "^26.0.1",
     "js-yaml": "^3.13.1",
     "lint-staged": "^9.5.0",
     "npm-run-all": "^4.1.5",
@@ -48,14 +56,13 @@
     "prop-types": "^15.6.2",
     "puppeteer": "^1.13.0",
     "raw-loader": "^1.0.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-helmet": "^5.2.0",
-    "regenerator-runtime": "^0.13.3",
     "remark": "^10.0.1",
-    "remark-cli": "^7.0.0",
-    "remark-frontmatter": "^1.3.2",
-    "remark-html": "^9.0.0",
+    "remark-cli": "^8.0.0",
+    "remark-frontmatter": "^2.0.0",
+    "remark-lint-heading-increment": "^2.0.1",
     "remark-react": "^5.0.1",
     "slugg": "^1.2.1",
     "tap-min": "^2.0.0",
