@@ -1,31 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import PageShell from './page_shell';
-import Feedback from '@mapbox/dr-ui/feedback';
-import constants from '../constants';
+import Feedback from './feedback';
 
 class MarkdownPageshell extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            user: undefined
-        };
-    }
-
-    componentDidMount() {
-        MapboxPageShell.afterUserCheck(() => {
-            // fetches username so we can identify them in segment
-            this.setState({
-                user: MapboxPageShell.getUser() || undefined
-            });
-        });
-    }
     render() {
         return (
             <PageShell {...this.props}>
-                <div className="prose">
-                    {this.props.location.pathname !==
-                        '/mapbox-gl-js/overview/' && (
+                <div className="prose mb18">
+                    {this.props.location.pathname !== '/mapbox-gl-js/api/' && (
                         <h1 className="txt-fancy">
                             {this.props.frontMatter.title}
                         </h1>
@@ -33,24 +16,24 @@ class MarkdownPageshell extends React.Component {
                     {this.props.children}
                 </div>
 
-                <div className="mt18">
-                    <Feedback
-                        site="Mapbox GL JS"
-                        location={this.props.location}
-                        user={this.state.user}
-                        webhook={constants.FORWARD_EVENT_WEBHOOK}
-                    />
-                </div>
+                {this.props.feedback && (
+                    <Feedback location={this.props.location} />
+                )}
             </PageShell>
         );
     }
 }
 
+MarkdownPageshell.defaultProps = {
+    feedback: true
+};
+
 MarkdownPageshell.propTypes = {
     frontMatter: PropTypes.object,
     location: PropTypes.object,
     meta: PropTypes.object,
-    children: PropTypes.node
+    children: PropTypes.node,
+    feedback: PropTypes.bool
 };
 
 export default MarkdownPageshell;
